Chang return type getStudentsQuantityByNames to Long

+++ src/main/java/ait/cohort46/student/controller/StudentController.java
@@ -47,7 +47,7 @@ public class StudentController {
     }
 
     @PostMapping ("/quantity/students")
-    public Integer getStudentsQuantityByNames(@RequestBody Set<String> names) {
+    public Long getStudentsQuantityByNames(@RequestBody Set<String> names) {
         return studentService.getStudentsQuantityByNames(names);
     }
