+++ src/main/java/ait/cohort46/student/service/StudentService.java
@@ -21,7 +21,7 @@ public interface StudentService {
 
     List<StudentDto> findStudentsByName(String name);
 
-    Integer getStudentsQuantityByNames(Set<String> names);
+    Long getStudentsQuantityByNames(Set<String> names);
 
     List<StudentDto> findStudentsByExamMinScore(String exam, Integer minScore);
 }
